refactor(jact-cli): extract CliFormatter, CitationFixer, and CLI wiring from JactCli (closes #29)
- src/cli.ts: new Commander entry point; all .command() registrations moved from jact.ts
- src/core/citationFixer.ts: CitationFixer class added with applyFixes(content, fixes, {dryRun}) method
- src/jact.ts: reduced from 1611→469 lines; JactCli now exports orchestration only, importable without activating Commander
- package.json: bin now points to dist/cli.js; all jact:* scripts updated
- test/unit/core/CitationFixer.test.ts: 8 unit tests for CitationFixer in isolation (no CLI spawn)
- All CLI integration test paths updated from dist/jact.js to dist/cli.js
