+++ test/story-validation.test.js
@@ -5,7 +5,7 @@ import { runCLI } from "./helpers/cli-runner.js";
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = dirname(__filename);
-const citationManagerPath = join(__dirname, "..", "dist", "jact.js");
+const citationManagerPath = join(__dirname, "..", "dist", "cli.js");
 
 describe("Story File Validation with Scope", () => {
 	it("should validate story file with mixed valid and broken citations", async () => {
