+++ test/warning-validation.test.js
@@ -5,7 +5,7 @@ import { runCLI } from "./helpers/cli-runner.js";
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = dirname(__filename);
-const citationManagerPath = join(__dirname, "..", "dist", "jact.js");
+const citationManagerPath = join(__dirname, "..", "dist", "cli.js");
 
 describe("Warning Status Validation Tests", () => {
 	it("should return warning status for cross-directory short filename citations resolved via file cache", async () => {
